+++ lib/helpers/index.js
@@ -4,10 +4,12 @@ const botConfig = require('./botConfig');
 const cache = require('./cache');
 const campaign = require('./campaign');
 const reportback = require('./reportback');
+const signup = require('./signup');
 
 module.exports = {
   botConfig,
   cache,
   campaign,
   reportback,
+  signup,
 };
