+++ lib/rogue.js
@@ -7,6 +7,7 @@ const logger = require('winston');
 const superagent = require('superagent');
 const ip = require('ip');
 
+const helpers = require('./helpers');
 const defaultConfig = require('../config/lib/rogue');
 
 const baseUri = defaultConfig.clientOptions.baseUri;
@@ -43,28 +44,10 @@ function executePost(endpoint, data) {
 }
 
 /**
- * @param {object} req
- * @return {object}
- */
-function getDefaultPayloadForReq(req) {
-  // @see https://github.com/DoSomething/rogue/blob/master/documentation/endpoints/signups.md#signups
-  const data = {
-    source: req.platform,
-    campaign_id: req.campaignId,
-    campaign_run_id: req.campaignRunId,
-    northstar_id: req.userId,
-  };
-  return data;
-}
-
-/**
- * @param {object} req
+ * @param {object} data
  * @return {Promise}
  */
-module.exports.createSignupForReq = function (req) {
-  const data = getDefaultPayloadForReq(req);
-  logger.debug('rogue.createSignupForReq', { data });
-
+module.exports.createSignup = function (data) {
   return executePost('signups', data);
 };
 
@@ -86,7 +69,7 @@ function fetchDataForImageUrl(url) {
  * @return {Promise}
  */
 module.exports.createPostForReq = function (req) {
-  const data = getDefaultPayloadForReq(req);
+  const data = helpers.signup.getDefaultCreatePayloadFromReq(req);
   const draft = req.signup.draft_reportback_submission;
 
   // @see https://github.com/DoSomething/rogue/blob/master/documentation/endpoints/posts.md
